Change traces for analysissessions to ADC. Also change the database field to IntegerField. Convert existing values in migrations. Requires newer SAPPHiRE. Update determination of arrival time.

+++ publicdb/coincidences/models.py
@@ -40,7 +40,7 @@ class Event(models.Model):
     station = models.ForeignKey(Station, models.CASCADE)
     pulseheights = ArrayField(models.IntegerField(), size=4)
     integrals = ArrayField(models.IntegerField(), size=4)
-    traces = ArrayField(ArrayField(models.FloatField()), size=4)
+    traces = ArrayField(ArrayField(models.IntegerField()), size=4)
 
     class Meta:
         ordering = ('date', 'time', 'nanoseconds', 'station')
